+++ src/app/(auth)/signup/complete/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import PrimaryButton from '@/components/PrimaryButton';
+import PrimaryButton from '@/components/primary-button';
 
 export default function SignUpComplete() {
   const router = useRouter();
